fix: FMin changes rseed arg during search
It was always unclear to me how randomness should be introduced to the suggest
algorithms, but now I think I understand how it should work.

1. Suggest() implementations should be stateless.
2. IDs should simply be labels, they should not actually affect search.

Therefore, the seed value passed to a suggest function should be different on
each evaluation, and that seed value should modulate all stochastic aspects of
the suggest() implementation.