Fix BSON import issue (#549)
A package `bson` has been added to the `setup.py` in commit 4dae46c4724d4796c6a83f146e8f336013ad9bdf,
because `pymongo` has been removed. The package is however not compatible
with the `bson` package from `pymongo` and has been causing
`ImportError`s.

Therefore we add all the imports of `bson` to a guarded block.
It should be safe to do so, because any code that needs the `bson`
package is also dependent on `pymongo`.